BitArray.cpp
90% of functions are done. Next step is to make >>= <<= and operator overloading.

+++ BitArray_lab1/BitArray_lab1/BitArray_lab1.h
@@ -8,7 +8,7 @@ class BitArray
 {
 private:
 	std::vector<char> bit_array;
-	size_t bit_count;
+	int bit_count;
 
 public:
 	BitArray();
